# backend/src/main/java/kuhcorp/orderbot/api/CommonDtos.java
package kuhcorp.orderbot.api;

import jakarta.validation.constraints.NotNull;
import lombok.NonNull;
import lombok.Value;

public class CommonDtos {

    @Value(staticConstructor = "of")
    public static class StringDto {
        @NotNull
        @NonNull
        String value;
    }

    @Value(staticConstructor = "of")
    public static class BooleanDto {
        @NotNull
        @NonNull
        Boolean value;
    }
}