# frontend/src/app/api/spec/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "OpenAPI documentation for My API",
    "title" : "My API",
    "version" : "59acdd34ea678de65cbe0a25f66df8cefdf5050b"
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost:8080"
  } ],
  "paths" : {
    "/api/p/template/manager/wip/{sessionId}/step/{stepId}" : {
      "delete" : {
        "operationId" : "deleteStep",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "sessionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        }, {
          "explode" : false,
          "in" : "path",
          "name" : "stepId",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "wipTemplateMng" ]
      },
      "get" : {
        "operationId" : "getStep",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "sessionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        }, {
          "explode" : false,
          "in" : "path",
          "name" : "stepId",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WipStepDetailRes"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "wipTemplateMng" ]
      },
      "put" : {
        "operationId" : "updateStep",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "sessionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        }, {
          "explode" : false,
          "in" : "path",
          "name" : "stepId",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/WipStepUpdateReq"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WipStepNodeData"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "wipTemplateMng" ]
      }
    },
    "/api/p/template/manager/wip/{sessionId}/step/{stepId}/location" : {
      "put" : {
        "operationId" : "updateStepLocation",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "sessionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        }, {
          "explode" : false,
          "in" : "path",
          "name" : "stepId",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/WipStepUpdatePositionReq"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "wipTemplateMng" ]
      }
    },
...
    "/api/p/template/manager/wip/{sessionId}/steps/connection" : {
      "post" : {
        "operationId" : "createConnection",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "sessionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/WipStepConnectionCreateReq"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StringDto"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "wipTemplateMng" ]
      }
    },
    "/api/p/template/manager/wip/{sessionId}/step" : {
      "post" : {
        "operationId" : "createStep",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "sessionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WipStepCreateData"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "wipTemplateMng" ]
      }
    },
    "/api/p/template/manager/wip/{sessionId}/complete" : {
      "post" : {
        "operationId" : "completeTemplate",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "sessionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "wipTemplateMng" ]
      }
    },
    "/api/p/template/manager/wip/session/{templateId}" : {
      "post" : {
        "operationId" : "getSession",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "templateId",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StringDto"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "wipTemplateMng" ]
      }
    },
...
    "/api/p/template/manager/wip/{sessionId}/validate" : {
      "get" : {
        "operationId" : "validateSteps",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "sessionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WipStepValidationRes"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "wipTemplateMng" ]
      }
    },
    "/api/p/template/manager/wip/{sessionId}/steps" : {
      "get" : {
        "operationId" : "getSteps",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "sessionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WipStepListRes"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "wipTemplateMng" ]
      }
    },
    "/api/p/template/manager/wip/{sessionId}/changed" : {
      "get" : {
        "operationId" : "isChanged",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "sessionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BooleanDto"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "wipTemplateMng" ]
      }
    },
...
    "/api/p/template/manager/wip/{sessionId}/steps/connection/{connectionId}" : {
      "delete" : {
        "operationId" : "deleteConnection",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "sessionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        }, {
          "explode" : false,
          "in" : "path",
          "name" : "connectionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "wipTemplateMng" ]
      }
    },
    "/api/p/template/manager/wip/{sessionId}/clear" : {
      "delete" : {
        "operationId" : "clearSession",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "sessionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StringDto"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "wipTemplateMng" ]
      }
    }
  },
  "components" : {
    "schemas" : {
      "TemplateStepPosition" : {
        "enum" : [ "FIRST", "MIDDLE", "LAST" ],
        "type" : "string"
      },
      "TemplateStepType" : {
        "enum" : [ "TEXT", "SELECT" ],
        "type" : "string"
      },
      "WipStepData" : {
        "properties" : {
          "type" : {
            "$ref" : "#/components/schemas/TemplateStepType"
          },
          "selectTypeData" : {
            "$ref" : "#/components/schemas/WipStepTypeSelect"
          }
        },
        "required" : [ "type" ],
        "type" : "object"
      },
      "WipStepTypeSelect" : {
        "properties" : {
          "options" : {
            "items" : {
              "$ref" : "#/components/schemas/WipStepTypeSelectOption"
            },
            "type" : "array"
          }
        },
        "required" : [ "options" ],
        "type" : "object"
      },
      "WipStepTypeSelectOption" : {
        "properties" : {
          "label" : {
            "type" : "string"
          },
          "value" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "WipStepUpdateReq" : {
        "properties" : {
          "title" : {
            "type" : "string"
          },
          "question" : {
            "maxLength" : 5000,
            "minLength" : 0,
            "type" : "string"
          },
          "orderPosition" : {
            "$ref" : "#/components/schemas/TemplateStepPosition"
          },
          "data" : {
            "$ref" : "#/components/schemas/WipStepData"
          }
        },
        "required" : [ "data", "orderPosition", "title" ],
        "type" : "object"
      },
      "WipStepListConnectionNode" : {
        "properties" : {
          "key" : {
            "type" : "string"
          },
          "label" : {
            "type" : "string"
          }
        },
        "required" : [ "key", "label" ],
        "type" : "object"
      },
      "WipStepNodeData" : {
        "properties" : {
          "title" : {
            "type" : "string"
          },
          "inputs" : {
            "items" : {
              "$ref" : "#/components/schemas/WipStepListConnectionNode"
            },
            "type" : "array"
          },
          "outputs" : {
            "items" : {
              "$ref" : "#/components/schemas/WipStepListConnectionNode"
            },
            "type" : "array"
          }
        },
        "required" : [ "inputs", "outputs", "title" ],
        "type" : "object"
      },
      "WipStepPosition" : {
        "properties" : {
          "x" : {
            "format" : "double",
            "type" : "number"
          },
          "y" : {
            "format" : "double",
            "type" : "number"
          }
        },
        "required" : [ "x", "y" ],
        "type" : "object"
      },
      "WipStepUpdatePositionReq" : {
        "properties" : {
          "position" : {
            "$ref" : "#/components/schemas/WipStepPosition"
          }
        },
        "required" : [ "position" ],
        "type" : "object"
      },
...
      "WipStepConnectionCreateReq" : {
        "properties" : {
          "sourceStepNumber" : {
            "type" : "string"
          },
          "targetStepNumber" : {
            "type" : "string"
          },
          "sourceOutput" : {
            "type" : "string"
          },
          "targetInput" : {
            "type" : "string"
          }
        },
        "required" : [ "sourceOutput", "sourceStepNumber", "targetInput", "targetStepNumber" ],
        "type" : "object"
      },
      "StringDto" : {
        "properties" : {
          "value" : {
            "type" : "string"
          }
        },
        "required" : [ "value" ],
        "type" : "object"
      },
      "WipStepCreateData" : {
        "properties" : {
          "stepNumber" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "question" : {
            "type" : "string"
          },
          "orderPosition" : {
            "$ref" : "#/components/schemas/TemplateStepPosition"
          },
          "data" : {
            "$ref" : "#/components/schemas/WipStepData"
          },
          "incomingConnections" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "outgoingConnections" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "gridPosition" : {
            "$ref" : "#/components/schemas/WipStepPosition"
          },
          "nodeData" : {
            "$ref" : "#/components/schemas/WipStepNodeData"
          }
        },
        "required" : [ "data", "gridPosition", "incomingConnections", "nodeData", "orderPosition", "outgoingConnections", "stepNumber", "title" ],
        "type" : "object"
      },
...
      "WipStepValidationRes" : {
        "properties" : {
          "isValid" : {
            "type" : "boolean"
          },
          "errorType" : {
            "enum" : [ "MISSING_STEPS", "MISSING_REQUIRED_FIELDS", "ONLY_ONE_FIRST_STEP_REQUIRED", "AT_LEAST_ONE_LAST_STEP_REQUIRED", "INVALID_STEP_CONNECTIONS" ],
            "type" : "string"
          },
          "errorMessage" : {
            "type" : "string"
          }
        },
        "required" : [ "isValid" ],
        "type" : "object"
      },
      "WipStepConnectionData" : {
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "sourceStepNumber" : {
            "type" : "string"
          },
          "targetStepNumber" : {
            "type" : "string"
          },
          "sourceOutput" : {
            "type" : "string"
          },
          "targetInput" : {
            "type" : "string"
          }
        },
        "required" : [ "id", "sourceOutput", "sourceStepNumber", "targetInput", "targetStepNumber" ],
        "type" : "object"
      },
      "WipStepListRes" : {
        "properties" : {
          "steps" : {
            "items" : {
              "$ref" : "#/components/schemas/WipStepListStep"
            },
            "maxItems" : 2147483647,
            "minItems" : 1,
            "type" : "array"
          },
          "connections" : {
            "items" : {
              "$ref" : "#/components/schemas/WipStepConnectionData"
            },
            "type" : "array"
          }
        },
        "required" : [ "connections", "steps" ],
        "type" : "object"
      },
      "WipStepListStep" : {
        "properties" : {
          "stepNumber" : {
            "type" : "string"
          },
          "orderPosition" : {
            "$ref" : "#/components/schemas/TemplateStepPosition"
          },
          "nodePosition" : {
            "$ref" : "#/components/schemas/WipStepPosition"
          },
          "nodeData" : {
            "$ref" : "#/components/schemas/WipStepNodeData"
          }
        },
        "required" : [ "nodeData", "nodePosition", "orderPosition", "stepNumber" ],
        "type" : "object"
      },
      "WipStepDetailRes" : {
        "properties" : {
          "stepNumber" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "question" : {
            "type" : "string"
          },
          "orderPosition" : {
            "$ref" : "#/components/schemas/TemplateStepPosition"
          },
          "data" : {
            "$ref" : "#/components/schemas/WipStepData"
          },
          "incomingConnections" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "outgoingConnections" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "required" : [ "data", "incomingConnections", "orderPosition", "outgoingConnections", "stepNumber", "title" ],
        "type" : "object"
      },
      "BooleanDto" : {
        "properties" : {
          "value" : {
            "type" : "boolean"
          }
        },
        "required" : [ "value" ],
        "type" : "object"
      },
...
    }
  }
}